feat(§BINANCE-P1): Phase 1 多市场地基——BrokerConfig + rules.binance + store 市场迁移 + qty REAL + 时段/风控 14 闸×3 市场矩阵 + verify 57（PLAN_BINANCE_MULTI_ASSET 2026-09-22）
- P1-a config/broker.go：BrokerConfig 接口 + QMTConfig 方法实现（CN 链零字段零 tag 变化，消费方泛型化留 Phase 2）
- P1-b config/binance.go + server/binance_config.go：rules.binance 出厂默认/normalize/validate + per-user KV + GET/POST /api/config/binance（脱敏回显镜像 qmt 口径）；发布闸顶层 Enabled=false 静态锁钉死
- P1-c store：orders/fills/shadow_orders/real_positions 等 market/currency 列 ALTER 循环 + real_positions 主键 (market,ts_code,user_id) 重建 + validTsCode 按市场放行 + 回报契约 market 透传（report_fields.json 双向锁）
- P1-d Qty int→float64 编译器驱动全波及 + migrateQtyReal 三表重建（REAL/INTEGER 双锁：REAL≥3 正锁 + INTEGER=2 豁免等值锁〔模拟盘簿 paper_trades / 史前库迁移过渡态〕）
- P1-e data/market_session.go：Session(market) 三市场时段抽象（CN 委托旧谓词零漂移 + CRYPTO 24h/UTC + US RTH/EXTENDED/24H 美东日历）；risk/gate.go 14 闸×3 市场矩阵——CN 专属四闸短路等值锁（=4）、min_notional CRYPTO 专属、lot_precision CRYPTO stepSize 容差/US 美分格 decimalPlaces、市场时区日键 marketToday（42 格矩阵测试全绿）
- §A2 契约三点锁同步：OrderRequest.market（omitempty，CN /order 载荷字节不变）落 order_fields.json + gateway.py ignored 集
- 测试基建：verify 57 段行为锁 8 组 + 静态锁 9 道；修 TestSweepOrdersStaleCancel 北京零点 §C1b 抢跑脆弱（改 14:00 锚点，00:01 实录红一次）
- 质量门：全量 go 47 ok + pytest 204 + verify_changes.sh 57 段 VERIFY_EXIT=0（CN 链零回归铁证）

+++ internal/engine/sell_unified_exec.go
@@ -20,7 +20,7 @@ import (
 
 // unifiedSellViews 把本轮裁决视图投影为 trading 层的统一卖出卡输入（mode=on 专用）。
 // sellable=T+1 可卖量（键=ts_code；缺 key=未知不锁定，与 §PROD-T1 放行口径一致）。
-func (e *Engine) unifiedSellViews(positions []store.RealPosition, verdicts []sellRoundVerdict, sellable map[string]int) []trading.UnifiedSellView {
+func (e *Engine) unifiedSellViews(positions []store.RealPosition, verdicts []sellRoundVerdict, sellable map[string]float64) []trading.UnifiedSellView {
 	if len(verdicts) == 0 {
 		return nil
 	}
@@ -90,7 +90,7 @@ func projectSellView(v *sellRoundVerdict) *trading.UnifiedSellView {
 }
 
 // t1LockedCodes §PROD-T1 口径：可卖量已知且 ≤0 的持仓代码集合（当日买入全锁）。
-func t1LockedCodes(positions []store.RealPosition, sellable map[string]int) map[string]bool {
+func t1LockedCodes(positions []store.RealPosition, sellable map[string]float64) map[string]bool {
 	out := map[string]bool{}
 	if len(sellable) == 0 {
 		return out
